Introduce XiaomiEuicc
This is a companion app for EuiccGoogle (SIM Manager), that basically
sends over the mapping of the SIM slots and sets proper SIM tray
illustration (props to Yuri for the lottie animations).

For the time being it only includes the mappings for one of the
Xiaoi 12T Pro models (22081212UG).

Change-Id: I09db5d6a9acbf05bf82803682572aeae54337c1b
Signed-off-by: Bruno Martins <bgcngm@gmail.com>
Signed-off-by: eidawon <projecteidawon@gmail.com>